refactor: pick컴포넌트를 분리한다
